feat(refractor/minestat): refractor the minestat module
Similarly, I have also refractored some bits of the
minestat module. When you're using setTimeout() and
setInterval(), it's always a good idea to keep a track
on the opened intervals and timeouts and clear them
(with clearTimeout and clearInterval) later.

Further, the indentation of the file was 2 ss instead
of 4 ss (like in your bot.js file). You should use either
ONE of those; not both.